// app/src/generated/models/expenseResponseCategory.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Category of the property-related expense
 */
export type ExpenseResponseCategory =
  (typeof ExpenseResponseCategory)[keyof typeof ExpenseResponseCategory];

export const ExpenseResponseCategory = {
  MAINTENANCE: 'MAINTENANCE',
  REPAIR: 'REPAIR',
  UTILITY: 'UTILITY',
  TAX: 'TAX',
  INSURANCE: 'INSURANCE',
  LEGAL: 'LEGAL',
  MARKETING: 'MARKETING',
  CLEANING: 'CLEANING',
  LANDSCAPING: 'LANDSCAPING',
  PROPERTY_MANAGEMENT: 'PROPERTY_MANAGEMENT',
  FEES: 'FEES',
  PROPERTY_TAX: 'PROPERTY_TAX',
  OTHER: 'OTHER',
} as const;